pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Nikita Renjov <n.renjov@cft.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mypy = "^1.10.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.14.0"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.19.2"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
